+++ src/Fill/FillMacros.py
@@ -11,11 +11,11 @@ def fill_o_arrow(
     ui_side.setCurrentIndex(c.attribute_values(c.ArrowSide).index(item["o_arrow"]["side"]))
     scene.skip_combobox_changes = False
 
-    ui_length_slider.setValue(10.0 * item["o_arrow"]["length"])
-    ui_length_spin.setValue(item["o_arrow"]["length"])
+    ui_length_slider.setValue(int(10.0 * item["o_arrow"]["length"]))
+    ui_length_spin.setValue(int(item["o_arrow"]["length"]))
 
-    ui_width_slider.setValue(10.0 * item["o_arrow"]["width"])
-    ui_width_spin.setValue(item["o_arrow"]["width"])
+    ui_width_slider.setValue(int(10.0 * item["o_arrow"]["width"]))
+    ui_width_spin.setValue(int(item["o_arrow"]["width"]))
 
     scene.skip_checkbox_changes = True
     ui_reversed.setChecked(item["o_arrow"]["reversed"])
@@ -32,11 +32,11 @@ def fill_d_arrow(
     ui_side.setCurrentIndex(c.attribute_values(c.ArrowSide).index(item["d_arrow"]["side"]))
     scene.skip_combobox_changes = False
 
-    ui_length_slider.setValue(10.0 * item["d_arrow"]["length"])
-    ui_length_spin.setValue(item["d_arrow"]["length"])
+    ui_length_slider.setValue(int(10.0 * item["d_arrow"]["length"]))
+    ui_length_spin.setValue(int(item["d_arrow"]["length"]))
 
-    ui_width_slider.setValue(10.0 * item["d_arrow"]["width"])
-    ui_width_spin.setValue(item["d_arrow"]["width"])
+    ui_width_slider.setValue(int(10.0 * item["d_arrow"]["width"]))
+    ui_width_spin.setValue(int(item["d_arrow"]["width"]))
 
     scene.skip_checkbox_changes = True
     ui_reversed.setChecked(item["d_arrow"]["reversed"])
@@ -56,11 +56,11 @@ def fill_colour(
     ui_mix_with.setCurrentIndex(colours.index(item["mix_with"]))
     scene.skip_combobox_changes = False
 
-    ui_mix_slider.setValue(item["mix_percent"])
-    ui_mix_spin.setValue(item["mix_percent"])
+    ui_mix_slider.setValue(int(item["mix_percent"]))
+    ui_mix_spin.setValue(int(item["mix_percent"]))
 
-    ui_strength_slider.setValue(item["strength"])
-    ui_strength_spin.setValue(item["strength"])
+    ui_strength_slider.setValue(int(item["strength"]))
+    ui_strength_spin.setValue(int(item["strength"]))
 
 def fill_fill_pattern(
         scene, item, ui_type,
@@ -74,20 +74,20 @@ def fill_fill_pattern(
     ui_type.setCurrentIndex(c.attribute_values(c.PatternType).index(item["type"]))
     scene.skip_combobox_changes = False
 
-    ui_distance_slider.setValue(10.0*item["distance"])
-    ui_distance_spin.setValue(item["distance"])
+    ui_distance_slider.setValue(int(10.0*item["distance"]))
+    ui_distance_spin.setValue(int(item["distance"]))
 
-    ui_size_slider.setValue(20.0*item["size"])
-    ui_size_spin.setValue(item["size"])
+    ui_size_slider.setValue(int(20.0*item["size"]))
+    ui_size_spin.setValue(int(item["size"]))
 
-    ui_rotation_slider.setValue(1.0/1.8*item["rotation"])
-    ui_rotation_spin.setValue(item["rotation"])
+    ui_rotation_slider.setValue(int(1.0/1.8*item["rotation"]))
+    ui_rotation_spin.setValue(int(item["rotation"]))
 
-    ui_xshift_slider.setValue(20.0*item["xshift"])
-    ui_xshift_spin.setValue(item["xshift"])
+    ui_xshift_slider.setValue(int(20.0*item["xshift"]))
+    ui_xshift_spin.setValue(int(item["xshift"]))
 
-    ui_yshift_slider.setValue(20.0*item["yshift"])
-    ui_yshift_spin.setValue(item["yshift"])
+    ui_yshift_slider.setValue(int(20.0*item["yshift"]))
+    ui_yshift_spin.setValue(int(item["yshift"]))
 
 def fill_decoration(
         scene, item, ui_type,
@@ -99,11 +99,11 @@ def fill_decoration(
     ui_type.setCurrentIndex(c.attribute_values(c.DecorationType).index(item["type"]))
     scene.skip_combobox_changes = False
 
-    ui_amplitude_slider.setValue(10.0*item["amplitude"])
-    ui_amplitude_spin.setValue(item["amplitude"])
+    ui_amplitude_slider.setValue(int(10.0*item["amplitude"]))
+    ui_amplitude_spin.setValue(int(item["amplitude"]))
 
-    ui_wavelength_slider.setValue(10.0*item["wavelength"])
-    ui_wavelength_spin.setValue(item["wavelength"])
+    ui_wavelength_slider.setValue(int(10.0*item["wavelength"]))
+    ui_wavelength_spin.setValue(int(item["wavelength"]))
 
     ui_text.setText(item["text"])
 
@@ -140,18 +140,18 @@ def fill_strategy(
     ui_bend_direction.setCurrentIndex(item["line"]["strategy"]["type"] != c.StrategyType.BENDED_LEFT)
     scene.skip_combobox_changes = False
 
-    ui_out_angle_slider.setValue(1.0/3.6*item["line"]["strategy"]["out_angle"])
-    ui_out_angle_spin.setValue(item["line"]["strategy"]["out_angle"])
+    ui_out_angle_slider.setValue(int(1.0/3.6*item["line"]["strategy"]["out_angle"]))
+    ui_out_angle_spin.setValue(int(item["line"]["strategy"]["out_angle"]))
 
-    ui_in_angle_slider.setValue(1.0/3.6*item["line"]["strategy"]["in_angle"])
-    ui_in_angle_spin.setValue(item["line"]["strategy"]["in_angle"])
+    ui_in_angle_slider.setValue(int(1.0/3.6*item["line"]["strategy"]["in_angle"]))
+    ui_in_angle_spin.setValue(int(item["line"]["strategy"]["in_angle"]))
 
-    ui_rounded_corners_slider.setValue(4.0*item["line"]["strategy"]["rounded_corners"])
-    ui_rounded_corners_spin.setValue(item["line"]["strategy"]["rounded_corners"])
+    ui_rounded_corners_slider.setValue(int(4.0*item["line"]["strategy"]["rounded_corners"]))
+    ui_rounded_corners_spin.setValue(int(item["line"]["strategy"]["rounded_corners"]))
 
-    ui_bend_angle_slider.setValue(1.0/3.6*item["line"]["strategy"]["bend_angle"])
-    ui_bend_angle_spin.setValue(item["line"]["strategy"]["bend_angle"])
+    ui_bend_angle_slider.setValue(int(1.0/3.6*item["line"]["strategy"]["bend_angle"]))
+    ui_bend_angle_spin.setValue(int(item["line"]["strategy"]["bend_angle"]))
 
-    ui_smooth_tension_slider.setValue(-10+50.0*item["line"]["strategy"]["smooth_tension"])
-    ui_smooth_tension_spin.setValue(item["line"]["strategy"]["smooth_tension"])
+    ui_smooth_tension_slider.setValue(int(-10+50.0*item["line"]["strategy"]["smooth_tension"]))
+    ui_smooth_tension_spin.setValue(int(item["line"]["strategy"]["smooth_tension"]))
